缺陷[#71125]: 常用功能数据库表设计
[描述说明]: 1.代码提交。2.问题修改.
[问题原因]:
[解决方法]:
[测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/user/service/SysUserService.java
@@ -772,6 +772,9 @@ public class SysUserService {
 
     @Transactional(rollbackFor = Exception.class)
     public int addCommonMenu(List<CommonMenu> list,Long userId){
+        if (list.isEmpty() || list.size() < 3){
+            throw new BusinessException(209,"常用功能至少保留三个！");
+        }
         //先删除 后增加
         commonMenuDao.deleteByUserId(userId);
         return commonMenuDao.batchAdd(list);
